+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.0 - Unreleased
+
+- Made the `app register` command to store application properties locally, so that they are easy to
+  use with other commands.
+- Added the `app requests` command for fetching the log of requests made by an application.
+- Small fixes and improvements.
+
 ## 0.1.0 - 2023-12-31
 
 The very first version with minimum functionality only allowing to authenticate as a user of Enable